chore: refactor packages in interface/presentation/application/domain
