+++ src/domain/aggregates/chat/message.ts
@@ -1,4 +1,4 @@
-import { User } from './user'
+import { User } from '../../user'
 
 export enum JoinNotification {
   JOIN_SESSION,
