+++ src/domain/aggregates/session/commands/utils.ts
