+++ src/domain/aggregates/video/events/videoEvents.ts
@@ -1,5 +1,5 @@
-import { SessionEvent, EventType } from '../../../../../domain/event/event'
-import { SessionReactions } from '../../../../../domain/reactions/sessionReactions'
+import { SessionEvent, EventType } from '../../../event/event'
+import { SessionReactions } from '../../../reactions/sessionReactions'
 
 export class VideoPlayedEvent implements SessionEvent {
   type: EventType
