+++ src/domain/aggregates/video/video.ts
@@ -1,6 +1,6 @@
-import { EventType } from '../../../../domain/event/event'
-import { EventBus } from '../../../../domain/event/eventBus'
-import { VideoReactions } from '../../../../domain/reactions/videoReactions'
+import { EventType } from '../../event/event'
+import { EventBus } from '../../event/eventBus'
+import { VideoReactions } from '../../reactions/videoReactions'
 import { User } from '../../user'
 import { isDeepEqual } from '../../utils'
 import { UserJoinedEvent, UserLeftSessionEvent } from '../session/events/sessionEvents'
