+++ src/domain/entity.ts
