+++ src/domain/reactions/chatReactions.ts
@@ -1,4 +1,9 @@
-import { NotificationMessage, TextMessage } from '../../application/session/message'
+import { NotificationMessage, TextMessage } from '../aggregates/chat/message'
+
+export enum ChatReactionType {
+  NOTIFICATION_MESSAGE = 'notificationMessage',
+  TEXT_MESSAGE = 'textMessage'
+}
 
 export interface ChatReactions {
   sendNotificationToSession(notificationMessage: NotificationMessage): void
