+++ src/domain/reactions/videoReactions.ts
@@ -1,4 +1,17 @@
-import { VideoState } from './reactions'
+export enum VideoReactionType {
+  VIDEO_STATE = 'videoState',
+  SYNCHRONIZE = 'synchronize'
+}
+
+export enum PlayState {
+  PAUSED = 'Paused',
+  PLAYING = 'Playing'
+}
+
+export interface VideoState {
+  timestamp: number
+  state: PlayState
+}
 
 export interface VideoReactions {
   retreiveVideoState(): Promise<VideoState>
